16. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -21,7 +21,7 @@ int main(void)
 		else if (i >= 1)
 			s = x + y;
 
-		printf("%ld", s);
+		printf("%lu", (long unsigned)s);
 
 		if (i == 98)
 			continue;
